Fixes seemingly random issue where messages would stop being received.
What was happening:

After we processed a chat message we removed it from the DOM.

This was normally not a problem unless a bot or an amdin would later
delete that chat message (rare, but happens reliably on popular
  channels). This would result in the pages own javascript attempting
to delete that DOM element (which we already deleted) -> this would
crash the page due to attempting to remove a node that didn't exist
anymore.

Solution:

Flag the processed DOM element instead using an internal id and filter
those messages from being processed further. The pages own javascript
will keep the chat log from becoming too large by deleting older
messages so we do not have to worry about size or memory issues.